requirement.txt
langchain==0.2.7
langchain-community==0.2.6
langchain-core==0.2.12
langchain-openai==0.1.7
langchain-text-splitters==0.2.0
pandas==2.2.2
pymssql==2.3.0
SQLAlchemy==2.0.30
openai==1.30.1
streamlit==1.36.0
streamlit-authenticator==0.2.3
python-dotenv
faiss-cpu==1.8.0.post1
extra-streamlit-components
azure-storage-blob
Werkzeug==2.3.6
Flask==2.3.2
 